requirements.txt
pickl
dill
dataclasses
numpy 
pandas
streamlit
google.generativeai
plotly.express
seaborn 
matplotlib
scikit-learn
